feat: improve search
